match odds preview

+++ src/main/java/kz/nmbet/betradar/dao/repository/GlCompetitorEntityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface GlCompetitorEntityRepository extends JpaRepository<GlCompetitorEntity, Integer> {
 
-	GlCompetitorEntity findByOutrightIdAndSuperIdAndTeamId(Integer id, Integer superId, Long teamId);
+	GlCompetitorEntity findByOutrightIdAndSuperIdAndTeamId(Integer id, Integer superId, Integer teamId);
 
-	GlCompetitorEntity findByMatchIdAndSuperIdAndTeamId(Integer id, Integer superTeamId, Long teamId);
+	GlCompetitorEntity findByMatchIdAndSuperIdAndTeamId(Integer id, Integer superTeamId, Integer teamId);
 }
